+++ CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add MetaLogger.Slicer to slice log entries.
+- Add body filtering to Tesla Middleware.
+- Slice log entries on Tesla Middleware.
 
 ## [1.3.1] - 2021-06-03
 ### Added
 - Custom replacements for filter patterns.
-
 
 ## [1.3.0] - 2021-06-02
 ### Added
@@ -18,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.2.0] - 2021-04-23
 ### Added
-- Add query params filtering to Tesla Middleware
+- Add query params filtering to Tesla Middleware.
 
 ### Changed
 - Ensure Tesla is loaded before defining MetaLogger middleware to avoid